requirements.txt
python-telegram-bot==13.15
requests==2.31.0
vnstock==3.2.2
pandas-ta==0.3.14b
ta==0.11.0
tenacity==9.0.0
